refactor(chat): decompose ChatInputBar (436→274) + drop orphan TakeoverDialog (Control v2 PR-6)
ChatInputBar was a 436-line god-component (>200 TS React smell). Extract its
logic into a focused hook and a state module, leaving a single-concern view:
- new input/useChatInput.ts — owns draft text, slash-popover, attachments,
  textarea ref + all keyboard/paste/slash/send handlers (behavior unchanged).
- new input/input-bar-state.ts — InputBarState + STATE_CONFIG + MODE_COMMANDS
  (one source of truth for placeholders / per-state flags). InputBarState is
  re-exported from ChatInputBar so consumers + tests are unaffected.
- ChatInputBar.tsx is now a thin view (274 lines) over the hook.

Also delete TakeoverDialog.tsx — orphaned since PR-3 swapped to the richer
TakeoverConfirmDialog (no production import remained).

Behavior-preserving: the existing 19 ChatInputBar tests (6 integration +
13 state-config) pass unchanged; web typecheck green.

Control v2, Pillar 4 (health).
